+++ database/migrations/2024_10_19_201331_create_table_referensi_users_jadwal_dinas.php
@@ -15,7 +15,7 @@ class CreateTableReferensiUsersJadwalDinas extends Migration
     {
         Schema::create('referensi_jadwal_users', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('pegawai_id')->comment('ID from Table Users');
+            $table->unsignedBigInteger('pegawai_id')->comment('ID from Table Users');
             $table->foreign('pegawai_id')->references('id')->on('users');
             $table->longText('staf')->nullable();
             $table->timestamps();
